requirements.txt
asgiref==3.10.0
Django==4.2.25
django-appconf==1.1.0
django-compressor==4.5.1
django-compressor-toolkit==0.6.0
django-image-uploader-widget==1.1.0
django-tinymce==5.0.0
mysqlclient==2.2.7
pillow==12.0.0
python-dotenv==1.1.1
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.3
